# Remove nif