--- a/rebar.config
+++ b/rebar.config
@@ -13,46 +13,19 @@
 ]}.
 
 {project_plugins, [
-    pc,
     rebar3_hex,
     rebar3_ex_doc
 ]}.
-
-{artifacts, [
-    "priv/jid.so"
-]}.
-
-{port_specs, [
-    {
-        ".*",
-        "priv/jid.so",
-        ["c_src/*.c"],
-        [{env, [{"CFLAGS", "$CFLAGS -std=c99 -O3 -Wall -Wextra -fPIC"}]}]
-     }
-]}.
-
-{provider_hooks, [
-    {post, [
-        {compile, {pc, compile}},
-        {clean, {pc, clean}}
-     ]}
- ]}.
 
 {profiles, [
     {test, [
         {cover_enabled, true},
         {cover_export_enabled, true},
         {deps, [{proper, "1.4.0"}]},
-        {plugins, [{rebar3_codecov, "0.6.0"}]},
-        {port_env,
-         [{"CFLAGS", "$CFLAGS -std=c99 -O2 -g -Wall -Wextra -fPIC --coverage"},
-          {"LDFLAGS", "$LDFLAGS --coverage"}
-         ]}
+        {plugins, [{rebar3_codecov, "0.6.0"}]}
      ]},
     {prod, [
-        {erl_opts, [deterministic]},
-        {port_env,
-         [{"CFLAGS", "$CFLAGS -std=c99 -O3 -Wall -Wextra -fPIC"}]}
+        {erl_opts, [deterministic]}
      ]}
 ]}.