🙂 
```
Name           ips        average  deviation         median         99th %
new         4.04 M        0.25 μs  ±4922.09%        0.21 μs        0.38 μs
old         0.60 M        1.66 μs   ±514.31%        1.63 μs        2.08 μs

Comparison: 
new         4.04 M
old         0.60 M - 6.69x slower +1.41 μs

Memory usage statistics:

Name    Memory usage
new            152 B
old            408 B - 2.68x memory usage +256 B
```

For the record, it is also faster in OTP26:

```
Operating System: macOS
CPU Information: Apple M3 Pro
Number of Available Cores: 12
Available memory: 36 GB
Elixir 1.16.3
Erlang 26.2.5
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 5 s
reduction time: 0 ns
parallel: 1
inputs: uuid
Estimated total run time: 24 s

Benchmarking new with input uuid ...
Benchmarking old with input uuid ...    
Calculating statistics...
Formatting results...

##### With input uuid #####
Name           ips        average  deviation         median         99th %
new         3.68 M        0.27 μs  ±5946.21%        0.25 μs        0.38 μs
old         0.53 M        1.90 μs   ±479.67%        1.83 μs        2.42 μs

Comparison: 
new         3.68 M
old         0.53 M - 7.01x slower +1.63 μs

Memory usage statistics:

Name    Memory usage
new            152 B
old            352 B - 2.32x memory usage +200 B

**All measurements for memory usage were the same**
```

On a RaspberryPi equivalent:
```
Compiling 1 file (-ex)
Generated benchs app
Operating System: Linux
CPU Information: ARMv8 Processor rev 4 (v81)
Number of Available Cores: 6
Available memory: 3.70 GB
Elixir 1.16.3
Erlang 26.2.5
JIT enabled: true
Benchmark suite executing with the following configuration: warmup: 2 s time: 5 s
memory time: 5 s reduction time: 0 ns parallel: 1 inputs: uuid
Estimated total run time: 24 s
Benchmarking new with input uuid Benchmarking old with input uuid
Calculating statistics...
Formatting results...
##### With input uuid #####
Name           ips        average  deviation         median         99th %
new         654.12 K        1.53 ms  ±983.47%        1.38 ms        2.33 ms
old         134.76 K       17.42 ms   ±311.80%        7.17 ms        8.50 ms

Comparison: 
new         654.12 K
old         134.76 K - 4.85x slower +5.89 ms

Memory usage statistics:

Name    Memory usage
new            152 B
old            360 B - 2.32x memory usage +208 B
```