@NelsonVides 2024-05-29
> That's interesting. Is this because of JIT or some other changes in the beam?

Lots of optimisations in binary matching, as well as type-based optimisations that allow the JIT to generate much better code. This algorithm is so so simple that the cost of calling a NIF makes a difference against the run of the actual NIF too 😄