@NelsonVides 2024-06-01
> Let's run tests on amd64 before merging it.

Results in a x86 Linux machine are satisfactory: 

OTP 26
```
Operating System: Linux
CPU Information: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
Number of Available Cores: 12
Available memory: 30.97 GB
Elixir 1.16.3
Erlang 26.2.5
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 5 s
reduction time: 0 ns
parallel: 1
inputs: uuid
Estimated total run time: 24 s

Benchmarking new with input uuid ...
Benchmarking old with input uuid ...
Calculating statistics...
Formatting results...

##### With input uuid #####
Name           ips        average  deviation         median         99th %
new         1.62 M        0.62 μs  ±5248.72%        0.51 μs        1.14 μs
old         0.35 M        2.87 μs   ±552.30%        2.60 μs        5.80 μs

Comparison: 
new         1.62 M
old         0.35 M - 4.65x slower +2.25 μs

Memory usage statistics:

Name    Memory usage
new            152 B
old            360 B - 2.37x memory usage +208 B

**All measurements for memory usage were the same**
```

And OTP 27:
```
Operating System: Linux
CPU Information: Intel(R) Core(TM) i9-8950HK CPU @ 2.90GHz
Number of Available Cores: 12
Available memory: 30.97 GB
Elixir 1.17.0-rc.0
Erlang 27.0
JIT enabled: true

Benchmark suite executing with the following configuration:
warmup: 2 s
time: 5 s
memory time: 5 s
reduction time: 0 ns
parallel: 1
inputs: uuid
Estimated total run time: 24 s

Benchmarking new with input uuid ...
Benchmarking old with input uuid ...
Calculating statistics...
Formatting results...

##### With input uuid #####
Name           ips        average  deviation         median         99th %
new         1.68 M        0.60 μs  ±4852.13%        0.46 μs        1.85 μs
old         0.34 M        2.95 μs   ±664.60%        2.71 μs        5.34 μs

Comparison: 
new         1.68 M
old         0.34 M - 4.95x slower +2.35 μs

Memory usage statistics:

Name    Memory usage
new            152 B
old            408 B - 2.68x memory usage +256 B

**All measurements for memory usage were the same**
```